feat(cloudflare): Add plugin for cloudflare pages
