+++ packages/cloudflare/src/handler.ts
@@ -1,23 +1,7 @@
-import type {
-  ExportedHandler,
-  ExportedHandlerFetchHandler,
-  IncomingRequestCfProperties,
-} from '@cloudflare/workers-types';
-import {
-  SEMANTIC_ATTRIBUTE_SENTRY_OP,
-  SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN,
-  SEMANTIC_ATTRIBUTE_SENTRY_SOURCE,
-  captureException,
-  continueTrace,
-  flush,
-  setHttpStatus,
-  startSpan,
-  withIsolationScope,
-} from '@sentry/core';
-import type { Options, Scope, SpanAttributes } from '@sentry/types';
-import { stripUrlQueryAndFragment, winterCGRequestToRequestData } from '@sentry/utils';
+import type { ExportedHandler, ExportedHandlerFetchHandler } from '@cloudflare/workers-types';
+import type { Options } from '@sentry/types';
 import { setAsyncLocalStorageAsyncContextStrategy } from './async';
-import { init } from './sdk';
+import { wrapRequestHandler } from './request';
 
 /**
  * Extract environment generic from exported handler.
@@ -47,70 +31,8 @@ export function withSentry<E extends ExportedHandler<any>>(
     handler.fetch = new Proxy(handler.fetch, {
       apply(target, thisArg, args: Parameters<ExportedHandlerFetchHandler<ExtractEnv<E>>>) {
         const [request, env, context] = args;
-        return withIsolationScope(isolationScope => {
-          const options = optionsCallback(env);
-          const client = init(options);
-          isolationScope.setClient(client);
-
-          const attributes: SpanAttributes = {
-            [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.http.cloudflare-worker',
-            [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'url',
-            [SEMANTIC_ATTRIBUTE_SENTRY_OP]: 'http.server',
-            ['http.request.method']: request.method,
-            ['url.full']: request.url,
-          };
-
-          const contentLength = request.headers.get('content-length');
-          if (contentLength) {
-            attributes['http.request.body.size'] = parseInt(contentLength, 10);
-          }
-
-          let pathname = '';
-          try {
-            const url = new URL(request.url);
-            pathname = url.pathname;
-            attributes['server.address'] = url.hostname;
-            attributes['url.scheme'] = url.protocol.replace(':', '');
-          } catch {
-            // skip
-          }
-
-          addRequest(isolationScope, request);
-          addCloudResourceContext(isolationScope);
-          if (request.cf) {
-            addCultureContext(isolationScope, request.cf);
-            attributes['network.protocol.name'] = request.cf.httpProtocol;
-          }
-
-          const routeName = `${request.method} ${pathname ? stripUrlQueryAndFragment(pathname) : '/'}`;
-
-          return continueTrace(
-            { sentryTrace: request.headers.get('sentry-trace') || '', baggage: request.headers.get('baggage') },
-            () => {
-              // Note: This span will not have a duration unless I/O happens in the handler. This is
-              // because of how the cloudflare workers runtime works.
-              // See: https://developers.cloudflare.com/workers/runtime-apis/performance/
-              return startSpan(
-                {
-                  name: routeName,
-                  attributes,
-                },
-                async span => {
-                  try {
-                    const res = await (target.apply(thisArg, args) as ReturnType<typeof target>);
-                    setHttpStatus(span, res.status);
-                    return res;
-                  } catch (e) {
-                    captureException(e, { mechanism: { handled: false, type: 'cloudflare' } });
-                    throw e;
-                  } finally {
-                    context.waitUntil(flush(2000));
-                  }
-                },
-              );
-            },
-          );
-        });
+        const options = optionsCallback(env);
+        return wrapRequestHandler({ options, request, context }, () => target.apply(thisArg, args));
       },
     });
 
@@ -119,20 +41,4 @@ export function withSentry<E extends ExportedHandler<any>>(
   }
 
   return handler;
-}
-
-function addCloudResourceContext(isolationScope: Scope): void {
-  isolationScope.setContext('cloud_resource', {
-    'cloud.provider': 'cloudflare',
-  });
-}
-
-function addCultureContext(isolationScope: Scope, cf: IncomingRequestCfProperties): void {
-  isolationScope.setContext('culture', {
-    timezone: cf.timezone,
-  });
-}
-
-function addRequest(isolationScope: Scope, request: Request): void {
-  isolationScope.setSDKProcessingMetadata({ request: winterCGRequestToRequestData(request) });
 }
