+++ packages/cloudflare/test/request.test.ts
@@ -1,0 +1,274 @@
+// Note: These tests run the handler in Node.js, which has some differences to the cloudflare workers runtime.
+// Although this is not ideal, this is the best we can do until we have a better way to test cloudflare workers.
+
+import { beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
+
+import * as SentryCore from '@sentry/core';
+import type { Event } from '@sentry/types';
+import { setAsyncLocalStorageAsyncContextStrategy } from '../src/async';
+import type { CloudflareOptions } from '../src/client';
+import { CloudflareClient } from '../src/client';
+import { wrapRequestHandler } from '../src/request';
+
+const MOCK_OPTIONS: CloudflareOptions = {
+  dsn: 'https://public@dsn.ingest.sentry.io/1337',
+};
+
+describe('withSentry', () => {
+  beforeAll(() => {
+    setAsyncLocalStorageAsyncContextStrategy();
+  });
+
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  test('passes through the response from the handler', async () => {
+    const response = new Response('test');
+    const result = await wrapRequestHandler(
+      { options: MOCK_OPTIONS, request: new Request('https://example.com'), context: createMockExecutionContext() },
+      () => response,
+    );
+    expect(result).toBe(response);
+  });
+
+  test('flushes the event after the handler is done using the cloudflare context.waitUntil', async () => {
+    const context = createMockExecutionContext();
+    await wrapRequestHandler(
+      { options: MOCK_OPTIONS, request: new Request('https://example.com'), context },
+      () => new Response('test'),
+    );
+
+    // eslint-disable-next-line @typescript-eslint/unbound-method
+    expect(context.waitUntil).toHaveBeenCalledTimes(1);
+    // eslint-disable-next-line @typescript-eslint/unbound-method
+    expect(context.waitUntil).toHaveBeenLastCalledWith(expect.any(Promise));
+  });
+
+  test('creates a cloudflare client and sets it on the handler', async () => {
+    const initAndBindSpy = vi.spyOn(SentryCore, 'initAndBind');
+    await wrapRequestHandler(
+      { options: MOCK_OPTIONS, request: new Request('https://example.com'), context: createMockExecutionContext() },
+      () => new Response('test'),
+    );
+
+    expect(initAndBindSpy).toHaveBeenCalledTimes(1);
+    expect(initAndBindSpy).toHaveBeenLastCalledWith(CloudflareClient, expect.any(Object));
+  });
+
+  describe('scope instrumentation', () => {
+    test('adds cloud resource context', async () => {
+      let sentryEvent: Event = {};
+      await wrapRequestHandler(
+        {
+          options: {
+            ...MOCK_OPTIONS,
+            beforeSend(event) {
+              sentryEvent = event;
+              return null;
+            },
+          },
+          request: new Request('https://example.com'),
+          context: createMockExecutionContext(),
+        },
+        () => {
+          SentryCore.captureMessage('cloud resource');
+          return new Response('test');
+        },
+      );
+
+      expect(sentryEvent.contexts?.cloud_resource).toEqual({ 'cloud.provider': 'cloudflare' });
+    });
+
+    test('adds request information', async () => {
+      let sentryEvent: Event = {};
+      await wrapRequestHandler(
+        {
+          options: {
+            ...MOCK_OPTIONS,
+            beforeSend(event) {
+              sentryEvent = event;
+              return null;
+            },
+          },
+          request: new Request('https://example.com'),
+          context: createMockExecutionContext(),
+        },
+        () => {
+          SentryCore.captureMessage('request');
+          return new Response('test');
+        },
+      );
+
+      expect(sentryEvent.sdkProcessingMetadata?.request).toEqual({
+        headers: {},
+        url: 'https://example.com/',
+        method: 'GET',
+      });
+    });
+
+    test('adds culture context', async () => {
+      const mockRequest = new Request('https://example.com') as any;
+      mockRequest.cf = {
+        timezone: 'UTC',
+      };
+
+      let sentryEvent: Event = {};
+      await wrapRequestHandler(
+        {
+          options: {
+            ...MOCK_OPTIONS,
+            beforeSend(event) {
+              sentryEvent = event;
+              return null;
+            },
+          },
+          request: mockRequest,
+          context: createMockExecutionContext(),
+        },
+        () => {
+          SentryCore.captureMessage('culture');
+          return new Response('test');
+        },
+      );
+
+      expect(sentryEvent.contexts?.culture).toEqual({ timezone: 'UTC' });
+    });
+  });
+
+  describe('error instrumentation', () => {
+    test('captures errors thrown by the handler', async () => {
+      const captureExceptionSpy = vi.spyOn(SentryCore, 'captureException');
+      const error = new Error('test');
+
+      expect(captureExceptionSpy).not.toHaveBeenCalled();
+
+      try {
+        await wrapRequestHandler(
+          { options: MOCK_OPTIONS, request: new Request('https://example.com'), context: createMockExecutionContext() },
+          () => {
+            throw error;
+          },
+        );
+      } catch {
+        // ignore
+      }
+
+      expect(captureExceptionSpy).toHaveBeenCalledTimes(1);
+      expect(captureExceptionSpy).toHaveBeenLastCalledWith(error, {
+        mechanism: { handled: false, type: 'cloudflare' },
+      });
+    });
+
+    test('re-throws the error after capturing', async () => {
+      const error = new Error('test');
+      let thrownError: Error | undefined;
+      try {
+        await wrapRequestHandler(
+          { options: MOCK_OPTIONS, request: new Request('https://example.com'), context: createMockExecutionContext() },
+          () => {
+            throw error;
+          },
+        );
+      } catch (e: any) {
+        thrownError = e;
+      }
+
+      expect(thrownError).toBe(error);
+    });
+  });
+
+  describe('tracing instrumentation', () => {
+    test('continues trace with sentry trace and baggage', async () => {
+      const mockRequest = new Request('https://example.com') as any;
+      mockRequest.headers.set('sentry-trace', '12312012123120121231201212312012-1121201211212012-1');
+      mockRequest.headers.set(
+        'baggage',
+        'sentry-release=2.1.12,sentry-public_key=public,sentry-trace_id=12312012123120121231201212312012,sentry-sample_rate=0.3232',
+      );
+
+      let sentryEvent: Event = {};
+      await wrapRequestHandler(
+        {
+          options: {
+            ...MOCK_OPTIONS,
+            tracesSampleRate: 0,
+            beforeSend(event) {
+              sentryEvent = event;
+              return null;
+            },
+          },
+          request: mockRequest,
+          context: createMockExecutionContext(),
+        },
+        () => {
+          SentryCore.captureMessage('sentry-trace');
+          return new Response('test');
+        },
+      );
+      expect(sentryEvent.contexts?.trace).toEqual({
+        parent_span_id: '1121201211212012',
+        span_id: expect.any(String),
+        trace_id: '12312012123120121231201212312012',
+      });
+    });
+
+    test('creates a span that wraps request handler', async () => {
+      const mockRequest = new Request('https://example.com') as any;
+      mockRequest.cf = {
+        httpProtocol: 'HTTP/1.1',
+      };
+      mockRequest.headers.set('content-length', '10');
+
+      let sentryEvent: Event = {};
+      await wrapRequestHandler(
+        {
+          options: {
+            ...MOCK_OPTIONS,
+            tracesSampleRate: 1,
+            beforeSendTransaction(event) {
+              sentryEvent = event;
+              return null;
+            },
+          },
+          request: mockRequest,
+          context: createMockExecutionContext(),
+        },
+        () => {
+          SentryCore.captureMessage('sentry-trace');
+          return new Response('test');
+        },
+      );
+
+      expect(sentryEvent.transaction).toEqual('GET /');
+      expect(sentryEvent.spans).toHaveLength(0);
+      expect(sentryEvent.contexts?.trace).toEqual({
+        data: {
+          'sentry.origin': 'auto.http.cloudflare',
+          'sentry.op': 'http.server',
+          'sentry.source': 'url',
+          'http.request.method': 'GET',
+          'url.full': 'https://example.com/',
+          'server.address': 'example.com',
+          'network.protocol.name': 'HTTP/1.1',
+          'url.scheme': 'https',
+          'sentry.sample_rate': 1,
+          'http.response.status_code': 200,
+          'http.request.body.size': 10,
+        },
+        op: 'http.server',
+        origin: 'auto.http.cloudflare',
+        span_id: expect.any(String),
+        status: 'ok',
+        trace_id: expect.any(String),
+      });
+    });
+  });
+});
+
+function createMockExecutionContext(): ExecutionContext {
+  return {
+    waitUntil: vi.fn(),
+    passThroughOnException: vi.fn(),
+  };
+}
